We can now dynamically switch between WinePickerApp and WinePickerAppE2E
We use ?mode=e2etest to identify e2e tests. When in this mode we load a
couple of extra JS files and set the ng-app to WinePickerAppE2E. This
allows us to use the e2e $httpBackend to return fake responses to
wine.com API calls. These calls are normally proxied by the server (see
WineApiController).